utils/load_tests/requirements.txt
locustio>=0.7.2
